+++ inst/include/vinecopulib/bicop/implementation/fit_controls.ipp
@@ -20,7 +20,8 @@ namespace vinecopulib {
 //!     `"linear"`, `"quadratic"`.
 //! @param nonparametric_mult a factor with which the smoothing parameters
 //!     are multiplied.
-//! @param selection_criterion the selection criterion (`"aic"` or `"bic"`).
+//! @param selection_criterion the selection criterion (`"loglik"`, `"aic"` 
+//!     or `"bic"`).
 //! @param preselect_families whether to exclude families before fitting
 //!     based on symmetry properties of the data.
 //! @param num_threads number of concurrent threads to use while fitting
@@ -95,8 +96,9 @@ FitControlsBicop::check_nonparametric_mult(double nonparametric_mult)
 inline void
 FitControlsBicop::check_selection_criterion(std::string selection_criterion)
 {
-    if (!tools_stl::is_member(selection_criterion, {"aic", "bic"})) {
-        throw std::runtime_error("selection_criterion should be aic or bic");
+    if (!tools_stl::is_member(selection_criterion, {"loglik", "aic", "bic"})) {
+        throw std::runtime_error(
+            "selection_criterion should be 'loglik', 'aic', or 'bic'");
     }
 }
 //! @}
@@ -185,7 +187,7 @@ inline void FitControlsBicop::set_num_threads(size_t num_threads)
 inline size_t FitControlsBicop::process_num_threads(size_t num_threads)
 {
     // use at least one thread
-    num_threads = std::max(num_threads, (size_t) 1);
+    num_threads = std::max(num_threads, static_cast<size_t>(1));
     // don't use more threads than supported by the system
     size_t max_threads = std::thread::hardware_concurrency();
     num_threads = std::min(num_threads, max_threads);
